+++ activesupport/lib/active_support/core_ext/range/overlap.rb
@@ -2,9 +2,11 @@
 
 class Range
   # Compare two ranges and see if they overlap each other
-  #  (1..5).overlaps?(4..6) # => true
-  #  (1..5).overlaps?(7..9) # => false
-  def overlaps?(other)
+  #  (1..5).overlap?(4..6) # => true
+  #  (1..5).overlap?(7..9) # => false
+  def overlap?(other)
     other.begin == self.begin || cover?(other.begin) || other.cover?(self.begin)
   end
+
+  alias :overlaps? :overlap?
 end